Enforce time limit

+++ TicTacToe/players.hpp
@@ -10,7 +10,7 @@ using namespace std;
 
 class HumanPlayer {
     public:
-        HumanPlayer(int t = 0)
+        HumanPlayer(double t = 0.0)
             : timeout(t)
         {}
 
@@ -30,28 +30,28 @@ class HumanPlayer {
             return tileNo;
         }
 
-        int getTimeout() const {
+        double getTimeout() const {
             return timeout;
         }
 
     private:
-        int timeout;
+        double timeout;
 };
 
 class StupidPlayer {
     public:
-        StupidPlayer(int t = 0)
+        StupidPlayer(double t = 0.0)
             : timeout(t)
         {}
         int getStrategy(const TicTacToe& state) {
             vector<int> possibleMoves = state.getAvailableMoves();
             return possibleMoves[0];
         }
-         int getTimeout() const {
+         double getTimeout() const {
             return timeout;
         }
     private:
-        int timeout;
+        double timeout;
 };
 
 #endif
